🚚 move: querykey 폴더 변경
