+++ src/components/eventMap/groups/SelectDetailGroup/SelectDetailGroup.tsx
@@ -3,7 +3,7 @@ import Icon from 'components/common/Icon/Icon';
 import Text from 'components/common/Text/Text';
 import SelectControlButton from 'components/eventMap/buttons/SelectControlButton/SelectControlButton';
 import SelectDetailBox from 'components/eventMap/selectboxes/SelectDetailBox/SelectDetailBox';
-import queryKeys from 'constants/queryKeys';
+import queryKeys from 'constants/queries/queryKeys';
 import { SelectStatus } from 'constants/selectBox';
 import {
   applicationAbleOptions,
