+++ src/constants/queries/queryKeys.ts
@@ -1,5 +1,5 @@
 import SortType from 'models/ledger/entity/SortType';
-import { FieldCode } from './interest';
+import { FieldCode } from '../interest';
 
 const DOMAIN = {
   USER: 'USER',
