+++ src/hooks/queries/event.ts
@@ -9,7 +9,7 @@ import {
   suspensionEvent,
 } from 'apis/eventInstance';
 import { FieldCode } from 'constants/interest';
-import queryKeys from 'constants/queryKeys';
+import queryKeys from 'constants/queries/queryKeys';
 import { CreateNewEventRequestDto } from 'models/event/request/CreateNewEventRequestDto';
 import EventSearchRequestDto from 'models/event/request/EventSearchRequestDto';
 import { EventSuspensionRequestDto } from 'models/event/request/EventSuspensionRequestDto';
