+++ src/screens/home/HomeScreen/HomeScreen.tsx
@@ -9,6 +9,7 @@ import useNavigator from 'hooks/navigator/useNavigator';
 import useInterestFields from 'hooks/interest/useInterestFields';
 import { usePersonalEventLists, useVogueEventLists } from 'hooks/queries/event';
 import { useMyInfo } from 'hooks/queries/user';
+import useResetQueries from 'hooks/queries/useResetQueries';
 import { useCallback, useEffect } from 'react';
 import { Image, ScrollView, TouchableOpacity, View } from 'react-native';
 import { foregroundListener, requestAlarmPermission } from 'services/fcm';
@@ -45,8 +46,10 @@ const HomeScreen = () => {
     stackNavigation.navigate(StackMenu.Alert);
   };
 
+  const { resetQueries } = useResetQueries();
+
   const foregroundListenerCallback = useCallback(() => {
-    foregroundListener();
+    foregroundListener({ resetQueries });
   }, []);
 
   useEffect(() => {
