+++ src/screens/userEvent/HostConsoleScreen/HostConsoleScreen.tsx
@@ -14,7 +14,7 @@ import {
 import MENT_HOST from 'constants/userEvent/host/hostMessage';
 import API_ERROR_MESSAGE from 'constants/errorMessage';
 import { EventDetailTabItem } from 'constants/eventDetail/eventDetailConstants';
-import queryKeys from 'constants/queryKeys';
+import queryKeys from 'constants/queries/queryKeys';
 import { ConsoleScreenLayout } from 'components/userEvent/host/layout';
 import FixedButton from 'components/common/FixedButton/FixedButton';
 import Spacing from 'components/common/Spacing/Spacing';
